Split and update tests
